+++ packages/ui/src/Slider/SliderArrow/index.tsx
@@ -5,6 +5,7 @@ import {ChevronLeftIcon, ChevronRightIcon} from '@eruditorgroup/profi-icons';
 import Button from '../../Button';
 
 import type {MouseEventHandler} from 'react';
+import type {ButtonProps} from '../../Button';
 
 import styles from './SliderArrow.module.scss';
 
@@ -14,6 +15,8 @@ type Props = {
   withFill?: boolean;
   background?: string;
   onClick: MouseEventHandler<HTMLButtonElement>;
+  arrowButton?: ButtonProps;
+  className?: string;
 };
 
 function compileGradient(
@@ -32,7 +35,9 @@ export default function SliderArrow({
   visible,
   onClick,
   background = '#fff',
+  arrowButton = {},
   withFill,
+  className,
 }: Props): JSX.Element {
   const Icon = direction === 'left' ? ChevronLeftIcon : ChevronRightIcon;
 
@@ -66,9 +71,15 @@ export default function SliderArrow({
         size="l"
         design="light"
         onClick={onClick}
-        className={cx(styles['arrow'], styles[`arrow_${direction}`], {
-          [styles['arrow_show']]: visible,
-        })}
+        {...arrowButton}
+        className={cx(
+          styles['arrow'],
+          styles[`arrow_${direction}`],
+          {
+            [styles['arrow_show']]: visible,
+          },
+          className,
+        )}
       >
         <Icon />
       </Button>